processor/requirements.txt
kafka-python
minio
opencv-python-headless
psycopg2-binary
ultralytics
